* commit 'c7a94331b6939f821790bdba22f97db012b45dd5':
  Changed current timestamp from 32bit to 64bit as part of modified platform_get_tick_ms api
  Removing unnecessary comment
  Reflecting review comments: timer_module init now does not pass parameter platform_Get_tick_ms can run much longer with 64bit variable storing accumulated ticks also accounted sleep time
  Adding modifications related to sleep functionality in sensor example
  Remving debug messages
  Working sleep functionality with lots of debug messages
  Correcting logical error regarding current timstamp
  Adding adjustment for current timestamp to account for the sleep interval
  Adding sleep functionality, changed main application to use sleep function instead of platform_delay api when appropriate. Resolved compilation error carried from previoud merge for when protocol selected is smsnas protocol

+++ include/cloud_comm/cloud_comm.h
@@ -297,7 +297,7 @@ cc_set_recv_result cc_set_recv_buffer(cc_buffer_desc *buf);
  * performs any pending activities and indicates when it next needs to be called
  * to service a time based event.
  */
-uint32_t cc_service_send_receive(uint32_t cur_ts);
+uint32_t cc_service_send_receive(uint64_t cur_ts);
 
 /**
  * \brief
